update website examples and merge with examples directory at the top-level
