+++ examples/.prettierrc
@@ -1,0 +1,4 @@
+{
+  "useTabs": false,
+  "tabWidth": 2
+}
