--- examples/async-generator/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Hello World</title>
-  </head>
-  <body>
-    <div id="mount"></div>
-    <script src="lib/index.js"></script>
-  </body>
-</html>
