--- examples/async-generator/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "crank-examples-hello-world",
-  "version": "0.1.0",
-  "private": true,
-  "files": [
-    "/lib"
-  ],
-  "scripts": {
-    "build": "rollup -c rollup.config.js"
-  },
-  "dependencies": {
-    "@bikeshaving/crank": "*"
-  },
-  "devDependencies": {
-    "rollup": "^2.3.4",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-typescript2": "^0.27.0",
-    "typescript": "^3.8.3"
-  }
-}
