--- examples/hackernews/README.md
@@ -1,6 +1,0 @@
-# Hacker News Clone with React
-
-## Running locally
-To start the demo application using the parcel bundler.
-
-- `npm start`
