--- examples/hackernews/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<title>Crank News</title>
-	</head>
-	<body>
-		<div id="app"></div>
-		<script src="./index.js"></script>
-	</body>
-</html>
