--- examples/hackernews/index.js
@@ -1,3 +1,0 @@
-import "core-js/stable";
-import "regenerator-runtime/runtime";
-import "./src/index";
