--- examples/hackernews/package.json
@@ -1,16 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "start": "parcel index.html"
-  },
-  "dependencies": {
-    "@bikeshaving/crank": "*",
-    "core-js": "^3.6.4",
-    "regenerator-runtime": "^0.13.5"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "parcel-bundler": "^1.12.4",
-    "typescript": "^3.8.3"
-  }
-}
