--- examples/hackernews/src/index.css
@@ -1,53 +1,0 @@
-.root {
-	font-family: sans-serif;
-	max-width: 800px;
-	margin: 0 auto;
-}
-
-.navbar {
-	display: flex;
-	justify-content: center;
-	margin: 30px 0;
-}
-
-.pager {
-	display: flex;
-	justify-content: center;
-}
-
-.story {
-	margin: 30px 0;
-}
-
-.item {
-	margin: 30px 0;
-}
-
-.meta {
-	color: #484848;
-}
-
-
-.comment {
-	margin: 30px 0;
-}
-
-.replies {
-	position: relative;
-	margin-left: 30px;
-}
-
-.expand {
-	position: relative;
-	top: -1px;
-	margin: 0;
-	padding: 0;
-	font-size: 0.8em;
-	vertical-align: botttom;
-	background-color: transparent;
-	border: none;
-}
-
-.expand:active {
-	color: black;
-}
