--- examples/letters/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "letters-example",
-  "private": true,
-  "version": "1.0.0",
-  "main": "index.html",
-  "scripts": {
-    "start": "parcel index.html --open",
-    "build": "parcel build index.html"
-  },
-  "eslint": {
-    "settings": {
-      "react": {
-        "pragma": "createElement"
-      }
-    }
-  },
-  "dependencies": {
-    "@bikeshaving/crank": "*"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.0",
-    "@babel/preset-react": "^7.13.13",
-    "parcel-bundler": "^1.12.5"
-  },
-  "browserslist": ["last 2 Chrome versions"],
-  "keywords": []
-}
