--- examples/simple/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Crank.js Sandbox</title>
-    <meta charset="UTF-8" />
-  </head>
-
-  <body>
-    <div id="app"></div>
-    <script src="src/index.tsx" type="module"></script>
-  </body>
-</html>
