--- examples/simple/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "a-simple-crank-component",
-  "private": true,
-  "version": "1.0.0",
-  "scripts": {
-    "start": "parcel index.html --open",
-    "build": "parcel build index.html"
-  },
-  "eslint": {
-    "settings": {
-      "react": {
-        "pragma": "createElement"
-      }
-    }
-  },
-  "dependencies": {
-    "@b9g/crank": "*",
-    "@babel/preset-react": "^7.18.6"
-  },
-  "devDependencies": {
-    "parcel": "^2.8.3",
-    "parcel-bundler": "^1.12.5"
-  },
-  "browserslist": [
-    "last 2 Chrome versions"
-  ],
-  "keywords": []
-}
