--- examples/todomvc-ts/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" data-framework="crank">
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>Crank • TodoMVC</title>
-		<link rel="stylesheet" href="node_modules/todomvc-common/base.css">
-		<link rel="stylesheet" href="node_modules/todomvc-app-css/index.css">
-	</head>
-	<body>
-		<section class="todoapp"></section>
-		<footer class="info">
-			<p>Double-click to edit a todo</p>
-			<p>Created by <a href="https://github.com/brainkim">Brian Kim</a></p>
-			<p>Part of <a href="http://todomvc.com">TodoMVC</a></p>
-		</footer>
-		<script src="node_modules/todomvc-common/base.js"></script>
-		<script src="./index.js"></script>
-	</body>
-</html>
