--- examples/todomvc-ts/package.json
@@ -1,19 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "start": "parcel index.html",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@bikeshaving/crank": "*",
-    "core-js": "^3.6.5",
-    "regenerator-runtime": "^0.13.5",
-    "todomvc-app-css": "2.2.0",
-    "todomvc-common": "^1.0.5"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "parcel-bundler": "^1.12.4",
-    "typescript": "^3.8.3"
-  }
-}
