--- examples/todomvc/README.md
@@ -1,10 +1,0 @@
-# Crank • [TodoMVC](http://todomvc.com)
-
-## Credit
-
-Created by [Brian Kim](https://github.com/brainkim)
-
-## Running locally
-To start the demo application using the parcel bundler.
-
-- `npm start`
