--- examples/todomvc/package.json
@@ -1,17 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "start": "parcel index.html"
-  },
-  "dependencies": {
-    "@bikeshaving/crank": "*",
-    "core-js": "^3.6.4",
-    "regenerator-runtime": "^0.13.5",
-    "todomvc-app-css": "2.2.0",
-    "todomvc-common": "^1.0.5"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "parcel-bundler": "^1.12.4"
-  }
-}
